pyproject.toml
[tool.poetry]
name = "treeserve"
version = "0.1.0"
description = ""
authors = ["gsingh <gds31.gagandeepsingh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
catboost = "^1.1.1"
lightgbm = "^3.3.5"
xgboost = "^1.7.3"
numpy = "^1.24.1"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"
boto3 = "^1.26.59"
structlog = "^22.3.0"
grpcio-health-checking = "^1.51.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

